+++ frontend_tests/zjsunit/zjquery_event.js
@@ -1,0 +1,15 @@
+"use strict";
+
+class FakeEvent {
+    constructor(type, props) {
+        if (!(this instanceof FakeEvent)) {
+            return new FakeEvent(type, props);
+        }
+        this.type = type;
+        Object.assign(this, props);
+    }
+    preventDefault() {}
+    stopPropagation() {}
+}
+
+module.exports = FakeEvent;
